changed find salary by gender output from int to double

+++ src/main/java/com/sparta/employeedatabase/entities/repository/SalaryRepository.java
@@ -32,5 +32,5 @@ public interface SalaryRepository extends JpaRepository<Salary, SalaryId> {
             + "from Salary s, Employee e \n"
             + "where s.empNo.id = e.id \n"
             + "and e.gender = :gender")
-    int findAvgSalaryByGender(String gender);
+    double findAvgSalaryByGender(String gender);
 }
